requirements.txt
absl-py==0.9.0
alembic==1.4.2
argon2-cffi==20.1.0
asn1crypto==0.24.0
astor==0.8.1
attrs==19.3.0
backcall==0.2.0
beautifulsoup4==4.7.1
bleach==3.1.5
certifi==2020.6.20
cffi==1.14.2
chardet==3.0.4
cliff==3.4.0
cloudpickle==1.1.1
cmaes==0.6.0
cmd2==1.3.3
colorama==0.4.3
colorlog==4.2.1
cryptography==2.1.4
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
docopt==0.6.2
docutils==0.16
entrypoints==0.3
funcsigs==1.0.2
furl==2.1.0
future==0.18.2
gast==0.2.2
gitdb==4.0.5
GitPython==3.1.7
google-pasta==0.2.0
grpcio==1.31.0
h5py==2.10.0
humanfriendly==8.2
idna==2.6
imageio==2.9.0
imgaug==0.4.0
importlib-metadata==1.7.0
ipdb==0.13.3
ipykernel==5.3.4
ipython==7.16.1
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.17.2
Jinja2==2.11.2
joblib==0.16.0
jsonpickle==1.4.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.6
jupyter-console==6.1.0
jupyter-core==4.6.3
Keras==2.4.3
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
keyring==10.6.0
keyrings.alt==3.0
kiwisolver==1.2.0
lockfile==0.12.2
luigi==3.0.1
lxml==4.5.2
Mako==1.1.3
Markdown==3.2.2
MarkupSafe==1.1.1
matplotlib==3.3.3
mistune==0.8.4
munch==2.5.0
nbconvert==5.6.1
nbformat==5.0.7
networkx==2.5
notebook==6.1.3
nteract-on-jupyter==2.1.3
numpy==1.19.1
opencv-python==4.4.0.42
opt-einsum==3.3.0
optuna==2.0.0
orderedmultidict==1.0.1
packaging==20.4
pandas==1.1.0
pandocfilters==1.4.2
parso==0.7.1
pathlib2==2.3.5
pbr==5.4.5
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.2.0
plotly==4.9.0
prettytable==0.7.2
prometheus-client==0.8.0
prompt-toolkit==3.0.6
protobuf==3.13.0
psutil==5.7.2
PTable==0.9.2
ptyprocess==0.6.0
py-cpuinfo==7.0.0
PyAudio==0.2.11
pycparser==2.20
pycrypto==2.6.1
pydub==0.24.1
Pygments==2.6.1
pygobject==3.26.1
PyJWT==1.7.1
pymongo==3.11.0
pyparsing==2.4.7
pyperclip==1.8.0
pyrsistent==0.16.0
python-daemon==2.2.4
python-dateutil==2.8.1
python-docx==0.8.10
python-editor==1.0.4
pytz==2020.1
PyWavelets==1.1.1
pyxdg==0.25
PyYAML==5.3.1
pyzmq==19.0.2
qtconsole==4.7.6
QtPy==1.9.0
requests==2.24.0
requests-file==1.5.1
retrying==1.3.3
sacred==0.8.1
scikit-image==0.17.2
scikit-learn==0.23.2
scipy==1.5.2
seaborn==0.11.0
SecretStorage==2.3.1
Send2Trash==1.5.0
Shapely==1.7.1
six==1.11.0
smmap==3.0.4
soupsieve==2.0.1
SQLAlchemy==1.3.19
ssh-import-id==5.7
stevedore==3.2.0
tensorboard==1.15.0
tensorflow-determinism==0.3.0
tensorflow-estimator==1.15.1
tensorflow-gan==1.0.0.dev0
tensorflow-gpu==1.15.3
tensorflow-probability==0.8.0
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
threadpoolctl==2.1.0
tifffile==2020.9.3
toml==0.10.1
torch==1.6.0
torchvision==0.7.0
tornado==5.1.1
tqdm==4.48.2
trains==0.16.0
traitlets==4.3.3
typing==3.7.4.3
urllib3==1.22
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
wrapt==1.12.1
zipp==3.1.0
